back_end/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.20

# Testing
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
